Added structure for basic TBA script interpreter. Handle the "%echo%" TBA script command.

+++ mind.h
@@ -60,7 +60,7 @@ private:
   int log;
 
   string script;
-  int spos;
+  size_t spos;
   Object *actor;
   };
